+++ pkg/admission/karydia/karydia.go
@@ -21,13 +21,11 @@ import (
 
 	"github.com/sirupsen/logrus"
 	"k8s.io/api/admission/v1beta1"
-	corev1 "k8s.io/api/core/v1"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
 
 	"github.com/karydia/karydia/pkg/k8sutil"
-	"github.com/karydia/karydia/pkg/k8sutil/scheme"
 )
 
 var resourcePod = metav1.GroupVersionResource{Group: "", Version: "v1", Resource: "pods"}
@@ -103,104 +101,6 @@ func (k *KarydiaAdmission) Admit(ar v1beta1.AdmissionReview, mutationAllowed boo
 	return k8sutil.AllowAdmissionResponse()
 }
 
-func (k *KarydiaAdmission) mutatePod(pod *corev1.Pod, ns *corev1.Namespace) *v1beta1.AdmissionResponse {
-	var patches []string
-
-	seccompProfile, annotated := ns.ObjectMeta.Annotations["karydia.gardener.cloud/seccompProfile"]
-	if annotated {
-		patches = mutatePodSeccompProfile(*pod, seccompProfile, patches)
-	}
-	return k8sutil.MutatingAdmissionResponse(patches)
-}
-
-func (k *KarydiaAdmission) validatePod(pod *corev1.Pod, ns *corev1.Namespace) *v1beta1.AdmissionResponse {
-	var validationErrors []string
-
-	seccompProfile, annotated := ns.ObjectMeta.Annotations["karydia.gardener.cloud/seccompProfile"]
-	if annotated {
-		validationErrors = validatePodSeccompProfile(*pod, seccompProfile, validationErrors)
-	}
-
-	return k8sutil.ValidatingAdmissionResponse(validationErrors)
-}
-
-func (k *KarydiaAdmission) mutateServiceAccount(sAcc *corev1.ServiceAccount, ns *corev1.Namespace) *v1beta1.AdmissionResponse {
-	var patches []string
-
-	automountServiceAccountToken, annotated := ns.ObjectMeta.Annotations["karydia.gardener.cloud/automountServiceAccountToken"]
-	if annotated {
-		patches = mutateServiceAccountTokenMount(*sAcc, automountServiceAccountToken, patches)
-	}
-
-	return k8sutil.MutatingAdmissionResponse(patches)
-}
-
-func (k *KarydiaAdmission) validateServiceAccount(sAcc *corev1.ServiceAccount, ns *corev1.Namespace) *v1beta1.AdmissionResponse {
-	var validationErrors []string
-
-	automountServiceAccountToken, annotated := ns.ObjectMeta.Annotations["karydia.gardener.cloud/automountServiceAccountToken"]
-	if annotated {
-		validationErrors = validateServiceAccountTokenMount(*sAcc, automountServiceAccountToken, validationErrors)
-	}
-
-	return k8sutil.ValidatingAdmissionResponse(validationErrors)
-}
-
-func validateServiceAccountTokenMount(sAcc corev1.ServiceAccount, nsAnnotation string, validationErrors []string) []string {
-	if nsAnnotation == "change-default" {
-		if automountServiceAccountTokenUndefined(&sAcc) && sAcc.Name == "default" {
-			validationErrors = append(validationErrors, "implicit automount of default service account token not allowed")
-		}
-	} else if nsAnnotation == "change-all" {
-		if automountServiceAccountTokenUndefined(&sAcc) {
-			validationErrors = append(validationErrors, "implicit automount of service account token not allowed")
-		}
-	}
-	return validationErrors
-}
-
-func mutateServiceAccountTokenMount(sAcc corev1.ServiceAccount, nsAnnotation string, patches []string) []string {
-	if nsAnnotation == "change-default" {
-		if automountServiceAccountTokenUndefined(&sAcc) && sAcc.Name == "default" {
-			patches = append(patches, fmt.Sprintf(`{"op": "add", "path": "/automountServiceAccountToken", "value": %s}`, "false"))
-		}
-	} else if nsAnnotation == "change-all" {
-		if automountServiceAccountTokenUndefined(&sAcc) {
-			patches = append(patches, fmt.Sprintf(`{"op": "add", "path": "/automountServiceAccountToken", "value": %s}`, "false"))
-		}
-	}
-	return patches
-}
-
-func validatePodSeccompProfile(pod corev1.Pod, nsAnnotation string, validationErrors []string) []string {
-	seccompPod, ok := pod.ObjectMeta.Annotations["seccomp.security.alpha.kubernetes.io/pod"]
-	if !ok || seccompPod != nsAnnotation {
-		validationErrorMsg := fmt.Sprintf("seccomp profile ('seccomp.security.alpha.kubernetes.io/pod') must be '%s'", nsAnnotation)
-		validationErrors = append(validationErrors, validationErrorMsg)
-	}
-	return validationErrors
-}
-
-func mutatePodSeccompProfile(pod corev1.Pod, nsAnnotation string, patches []string) []string {
-	_, ok := pod.ObjectMeta.Annotations["seccomp.security.alpha.kubernetes.io/pod"]
-	if !ok {
-		if len(pod.ObjectMeta.Annotations) == 0 {
-			// If no annotation object exists yet, we have
-			// to create it. Otherwise we will encounter
-			// the following error:
-			// 'jsonpatch add operation does not apply: doc is missing path: "/metadata/annotations..."'
-			patches = append(patches, fmt.Sprintf(`{"op": "add", "path": "/metadata/annotations", "value": {"%s": "%s"}}`, "seccomp.security.alpha.kubernetes.io/pod", nsAnnotation))
-		} else {
-			patches = append(patches, fmt.Sprintf(`{"op": "add", "path": "/metadata/annotations/%s", "value": "%s"}`, "seccomp.security.alpha.kubernetes.io~1pod", nsAnnotation))
-		}
-	}
-	return patches
-}
-
-func automountServiceAccountTokenUndefined(sAcc *corev1.ServiceAccount) bool {
-	return sAcc.AutomountServiceAccountToken == nil
-}
-
 func (k *KarydiaAdmission) getNamespaceFromAdmissionRequest(ar v1beta1.AdmissionRequest) (*v1.Namespace, error) {
 	namespaceRequest := ar.Namespace
 	if namespaceRequest == "" {
@@ -213,25 +113,6 @@ func (k *KarydiaAdmission) getNamespaceFromAdmissionRequest(ar v1beta1.Admission
 		return nil, e
 	}
 	return namespace, nil
-}
-
-/* Utility functions to decode raw resources into objects */
-func decodePod(raw []byte) (*corev1.Pod, error) {
-	pod := &corev1.Pod{}
-	deserializer := scheme.Codecs.UniversalDeserializer()
-	if _, _, err := deserializer.Decode(raw, nil, pod); err != nil {
-		return nil, err
-	}
-	return pod, nil
-}
-
-func decodeServiceAccount(raw []byte) (*corev1.ServiceAccount, error) {
-	sAcc := &corev1.ServiceAccount{}
-	deserializer := scheme.Codecs.UniversalDeserializer()
-	if _, _, err := deserializer.Decode(raw, nil, sAcc); err != nil {
-		return nil, err
-	}
-	return sAcc, nil
 }
 
 func shouldIgnoreEvent(ar v1beta1.AdmissionReview) bool {
