+++ pkg/admission/karydia/pod_admission.go
@@ -1,0 +1,83 @@
+// Copyright (C) 2019 SAP SE or an SAP affiliate company. All rights reserved.
+// This file is licensed under the Apache Software License, v. 2 except as
+// noted otherwise in the LICENSE file.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package karydia
+
+import (
+	"fmt"
+
+	"k8s.io/api/admission/v1beta1"
+	corev1 "k8s.io/api/core/v1"
+
+	"github.com/karydia/karydia/pkg/k8sutil"
+	"github.com/karydia/karydia/pkg/k8sutil/scheme"
+)
+
+func (k *KarydiaAdmission) mutatePod(pod *corev1.Pod, ns *corev1.Namespace) *v1beta1.AdmissionResponse {
+	var patches []string
+
+	seccompProfile, annotated := ns.ObjectMeta.Annotations["karydia.gardener.cloud/seccompProfile"]
+	if annotated {
+		patches = mutatePodSeccompProfile(*pod, seccompProfile, patches)
+	}
+	return k8sutil.MutatingAdmissionResponse(patches)
+}
+
+func (k *KarydiaAdmission) validatePod(pod *corev1.Pod, ns *corev1.Namespace) *v1beta1.AdmissionResponse {
+	var validationErrors []string
+
+	seccompProfile, annotated := ns.ObjectMeta.Annotations["karydia.gardener.cloud/seccompProfile"]
+	if annotated {
+		validationErrors = validatePodSeccompProfile(*pod, seccompProfile, validationErrors)
+	}
+
+	return k8sutil.ValidatingAdmissionResponse(validationErrors)
+}
+
+func validatePodSeccompProfile(pod corev1.Pod, nsAnnotation string, validationErrors []string) []string {
+	seccompPod, ok := pod.ObjectMeta.Annotations["seccomp.security.alpha.kubernetes.io/pod"]
+	if !ok || seccompPod != nsAnnotation {
+		validationErrorMsg := fmt.Sprintf("seccomp profile ('seccomp.security.alpha.kubernetes.io/pod') must be '%s'", nsAnnotation)
+		validationErrors = append(validationErrors, validationErrorMsg)
+	}
+	return validationErrors
+}
+
+func mutatePodSeccompProfile(pod corev1.Pod, nsAnnotation string, patches []string) []string {
+	_, ok := pod.ObjectMeta.Annotations["seccomp.security.alpha.kubernetes.io/pod"]
+	if !ok {
+		if len(pod.ObjectMeta.Annotations) == 0 {
+			// If no annotation object exists yet, we have
+			// to create it. Otherwise we will encounter
+			// the following error:
+			// 'jsonpatch add operation does not apply: doc is missing path: "/metadata/annotations..."'
+			patches = append(patches, fmt.Sprintf(`{"op": "add", "path": "/metadata/annotations", "value": {"%s": "%s"}}`, "seccomp.security.alpha.kubernetes.io/pod", nsAnnotation))
+		} else {
+			patches = append(patches, fmt.Sprintf(`{"op": "add", "path": "/metadata/annotations/%s", "value": "%s"}`, "seccomp.security.alpha.kubernetes.io~1pod", nsAnnotation))
+		}
+	}
+	return patches
+}
+
+/* Utility functions to decode raw resources into objects */
+func decodePod(raw []byte) (*corev1.Pod, error) {
+	pod := &corev1.Pod{}
+	deserializer := scheme.Codecs.UniversalDeserializer()
+	if _, _, err := deserializer.Decode(raw, nil, pod); err != nil {
+		return nil, err
+	}
+	return pod, nil
+}
